+++ src/core/backend/avx512/circle.rs
@@ -5,10 +5,10 @@ use super::m31::PackedBaseField;
 use super::{as_cpu_vec, AVX512Backend, VECS_LOG_SIZE};
 use crate::core::backend::avx512::fft::rfft;
 use crate::core::backend::avx512::BaseFieldVec;
-use crate::core::backend::CPUBackend;
+use crate::core::backend::{CPUBackend, Col};
 use crate::core::circle::{CirclePoint, Coset};
 use crate::core::fields::m31::BaseField;
-use crate::core::fields::{Col, ExtensionOf, FieldExpOps};
+use crate::core::fields::{ExtensionOf, FieldExpOps};
 use crate::core::poly::circle::{
     CanonicCoset, CircleDomain, CircleEvaluation, CirclePoly, PolyOps,
 };
@@ -178,8 +178,8 @@ impl PolyOps for AVX512Backend {
 mod tests {
     use crate::core::backend::avx512::fft::{CACHED_FFT_LOG_SIZE, MIN_FFT_LOG_SIZE};
     use crate::core::backend::avx512::AVX512Backend;
+    use crate::core::backend::Column;
     use crate::core::fields::m31::BaseField;
-    use crate::core::fields::Column;
     use crate::core::poly::circle::{CanonicCoset, CircleDomain, CircleEvaluation, CirclePoly};
     use crate::core::poly::{BitReversedOrder, NaturalOrder};
